requirements.txt
Django~=2.2.4
pytest-django==3.8.0
django-debug-toolbar==2.2
python-slugify==4.0.0
Pillow==7.0.0